requirements.txt
aiohttp
aiosqlite
discord.py
python-dotenv
openskill
nextcord
sqlalchemy[asyncio]
asyncpg
alembic